+++ src/Main/Features/General/AutoFollowWindow.cpp
