+++ src/Main/Features/General/AutoFollowWindow.h
