+++ src/Main/Features/General/CancelActionWindow.cpp
