+++ src/Main/Features/Uw/EmoWindow.cpp
@@ -19,7 +19,7 @@
 #include <GWCA/Managers/PartyMgr.h>
 #include <GWCA/Packets/Opcodes.h>
 
-#include <Actions.h>
+#include <ActionsUw.h>
 #include <Base/HelperBox.h>
 #include <GuiUtils.h>
 #include <Helper.h>
