+++ src/Main/Features/Uw/EmoWindow.h
@@ -7,7 +7,7 @@
 #include <GWCA/Packets/StoC.h>
 #include <GWCA/Utilities/Hook.h>
 
-#include <Actions.h>
+#include <ActionsUw.h>
 #include <AgentData.h>
 #include <Base/HelperBoxWindow.h>
 #include <GuiUtils.h>
