+++ src/Main/Features/Uw/TerraWindow.cpp
@@ -11,7 +11,7 @@
 #include <GWCA/GameEntities/Player.h>
 #include <GWCA/Managers/PartyMgr.h>
 
-#include <Actions.h>
+#include <ActionsUw.h>
 #include <Base/HelperBox.h>
 #include <GuiUtils.h>
 #include <HelperAgents.h>
